// src/components/spacex/types.ts
export interface FirstStage {
  cores: Core[];
}
// ...
export interface SecondStage {
  block?: number;
  payloads: Payload[];
}

export interface Fairings {
  reused?: boolean;
  recovery_attempt?: boolean;
  recovered?: boolean;
  ship: string;
}

export interface Rocket {
  rocket_id: string;
  rocket_name: string;
  rocket_type: string;
  first_stage: FirstStage;
  second_stage: SecondStage;
  fairings: Fairings;
}

export interface Telemetry {
  flight_club: string;
}

export interface Reuse {
  core: boolean;
  side_core1: boolean;
  side_core2: boolean;
  fairings: boolean;
  capsule: boolean;
}

export interface LaunchSite {
  site_id: string;
  site_name: string;
  site_name_long: string;
}

export interface LaunchFailureDetails {
  time: number;
  altitude?: number;
  reason: string;
}

export interface Links {
  mission_patch: string;
  mission_patch_small: string;
  reddit_campaign: string;
  reddit_launch: string;
  reddit_recovery: string;
  reddit_media: string;
  presskit: string;
  article_link: string;
  wikipedia: string;
  video_link: string;
  youtube_id: string;
  flickr_images: string[];
}

export interface Timeline {
  webcast_liftoff?: number;
  go_for_prop_loading?: number;
  rp1_loading?: number;
  stage1_lox_loading?: number;
  stage2_lox_loading?: number;
  engine_chill?: number;
  prelaunch_checks?: number;
  propellant_pressurization?: number;
  go_for_launch?: number;
  ignition?: number;
  liftoff?: number;
  maxq?: number;
  meco?: number;
  stage_sep?: number;
  second_stage_ignition?: number;
  dragon_separation?: number;
  dragon_solar_deploy?: number;
  dragon_bay_door_deploy?: number;
  fairing_deploy?: number;
  payload_deploy?: number;
  second_stage_restart?: number;
  webcast_launch?: number;
  payload_deploy_1?: number;
  payload_deploy_2?: number;
  first_stage_boostback_burn?: number;
  first_stage_entry_burn?: number;
  first_stage_landing?: number;
  beco?: number;
  side_core_sep?: number;
  side_core_boostback?: number;
  center_stage_sep?: number;
  center_core_boostback?: number;
  side_core_entry_burn?: number;
  center_core_entry_burn?: number;
  side_core_landing?: number;
  center_core_landing?: number;
  first_stage_landing_burn?: number;
  stage1_rp1_loading?: number;
  stage2_rp1_loading?: number;
}

export interface Launch {
  flight_number: number;
  mission_name: string;
  mission_id: string[];
  upcoming: boolean;
  launch_year: string;
  launch_date_unix: number;
  launch_date_utc: Date;
  launch_date_local: Date;
  is_tentative: boolean;
  tentative_max_precision: string;
  tbd: boolean;
  launch_window?: number;
  rocket: Rocket;
  ships: string[];
  telemetry: Telemetry;
  reuse: Reuse;
  launch_site: LaunchSite;
  launch_success?: boolean;
  launch_failure_details: LaunchFailureDetails;
  links: Links;
  details: string;
  static_fire_date_utc?: Date;
  static_fire_date_unix?: number;
  timeline: Timeline;
  crew: any[];
  last_date_update?: Date;
  last_ll_launch_date?: Date;
  last_ll_update?: Date;
  last_wiki_launch_date?: Date;
  last_wiki_revision: string;
  last_wiki_update?: Date;
  launch_date_source: string;
}